Implemented the event listener for key presses on the game and started on the board
As stated in the summary, the event listener for the game board has been implmented so all the user has to do is to type the word and it will only accept the following keys for the game A-Z, a-z, enter, and backspace.  Lastly, I hardcoded a table so we can get an idea of how we could have to board look.
 The table being hard coded is just temporary, and I added some inline styles for how we could select specific cells in the table.  I am however unsure if this will be the way we do it, I am just brainstorming some ideas.