Validation of the word completed and fixed Issue with navigating back to home screen after logging in
Made the game validate the word supplied to see if it is a word and it is on the board.  Also, fixed an issue with the navigation for the client when they login and then go to the URL bar and type "localhost:3000" and navigating back to the home screen.