Modified the server and client logic for the logic so I can start on the selection process for the word on the client's side
Moved the game board to the session variable (I think this is what we want to do for this?).
Cleaned up the output for determining if the word is valid on the server.
Modified the client in how it displays the board so it has the same board as the the server.  This is because it will simplify the logic for selecting the cells when the user types a word.